many more type castings. compiles

+++ protocol/contracts/beanstalk/init/InitBipSeedGauge.sol
@@ -120,7 +120,7 @@ contract InitBipSeedGauge is Weather, InitWhitelistStatuses {
         emit BeanToMaxLpGpPerBdvRatioChange(
             s.season.current,
             type(uint256).max,
-            int80(s.seedGauge.beanToMaxLpGpPerBdvRatio)
+            int80(int128(s.seedGauge.beanToMaxLpGpPerBdvRatio))
         );
 
         emit LibGauge.UpdateAverageStalkPerBdvPerSeason(
