+++ protocol/contracts/mocks/mockFacets/MockAdminFacet.sol
@@ -59,7 +59,7 @@ contract MockAdminFacet is Sun {
 
     function updateStart() private {
         SeasonFacet sf = SeasonFacet(address(this));
-        int256 sa = s.season.current - sf.seasonTime();
+        int256 sa = int256(uint256(s.season.current - sf.seasonTime()));
         if (sa >= 0) s.season.start -= 3600 * (uint256(sa) + 1);
     }
